Simplify, simplify
Calling into the twitch-dl app fell on its face the first day I put it into
production. :-( Instead I'm just going to call it from the command-line, which
seems to fare better. Also,l I installed it with "pipx" which might do a
better job of ensuring it has access to all the resources it needs.